backend/requirements.txt
# Backend Dependencies

# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
alembic==1.13.1

# Image Processing
opencv-python==4.9.0.80
numpy==1.26.3
pytesseract==0.3.10
Pillow==10.2.0
scikit-image==0.22.0

# 3D Processing
trimesh==4.1.4
numpy-stl==3.1.1

# Graph & Algorithms
networkx==3.2.1
scipy==1.12.0
sknw>=0.3

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# Validation
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# Utils
python-dotenv==1.0.0
aiofiles==23.2.1
httpx==0.26.0
shapely>=2.0.0
mapbox-earcut>=1.0.0
